fix: trying to get links to update correctly
